Previously installed HUD is now deleted
When switching between installer variants, the HUD install directory name varies. Previously a situation could arise where the a previous HUD directory was still present and populated and a new HUD directory (with a different name) would be created, ignoring the older directory.

If the HUD being installed is different from the already installed HUD, the installed HUD will be purged.